[refactor] Font Size Item Component 리팩토링
