+++ client/src/lib/constants/constants.ts
@@ -1,1 +1,8 @@
 export const TAB_LIMIT: number = 4;
+export const FONT_SIZE_LISTS = [
+  { size: '10', checked: false },
+  { size: '15', checked: true },
+  { size: '20', checked: false },
+  { size: '25', checked: false },
+  { size: '30', checked: false },
+];
